arvyn-server/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asgi]
langgraph
playwright
google-genai
pydantic
python-dotenv